requirements.txt
torch
torchvision
pillow
numpy
opencv-python
pandas
matplotlib
vietocr
paddleocr
pdf2image
paddlepaddle